Combine FloorItems and FloorItemFactory
