+++ game/src/main/kotlin/world/gregs/voidps/Main.kt
@@ -27,7 +27,6 @@ import world.gregs.voidps.engine.entity.character.player.login.LoginQueue
 import world.gregs.voidps.engine.entity.character.player.login.loginQueueModule
 import world.gregs.voidps.engine.entity.character.update.visualUpdatingModule
 import world.gregs.voidps.engine.entity.definition.detailsModule
-import world.gregs.voidps.engine.entity.item.floorItemModule
 import world.gregs.voidps.engine.entity.item.floorItemSpawnModule
 import world.gregs.voidps.engine.entity.list.entityListModule
 import world.gregs.voidps.engine.entity.obj.customObjectModule
@@ -178,7 +177,6 @@ object Main {
                 npcLoaderModule,
                 npcSpawnModule,
                 stairsModule,
-                floorItemModule,
                 floorItemSpawnModule,
                 musicModule,
                 areasModule
